requirements.txt
streamlit==1.26.0
torch>=1.13.0
transformers>=4.30.0
sentence-transformers==2.2.2
pillow>=9.0.0
requests>=2.28.0
pymilvus==2.2.1
nest-asyncio==1.5.6
torchvision>=0.14.0
